Rename model package to models.
